Introduce framework to compare different predictions
The AgePredictorComparison allows to specify multiple hyperparameter and
preprocessing configurations and compare the prediction results directly.

It supports a comparison_config which includes a list of predictor
configurations which are tested seqentually, as well as a
grid_comparison_config for which every configuration is tested in
combination with every comparison_config resulting in m x n prediction
results.