Introduce neighborbood cross-validation method
A neighborbood corresponds to a group of neighboring street blocks with
a maximal diameter of 1km which have been aggregated by agglomerative
clustering.

Neighborbood cross-validation is a compromise between city and street
based cross-validation with the goal to reduce spatial autocorrelation
between chunks sufficiently to prevent overly optimistic prediction
result while not making the prediction problem as difficult as
generalizing to distant cities. Neighborboods with a maximal diameter of
1km have shown to half the spatial autocorrelation and reduce Moran's I
to below 0.1.

The Haversine distance measure is needed to determine the metric distance
between two pairs of coordinates.