Consider unknown columns as aux vars and not as features
Previously, columns which haven't explicitly been labeled as auxiliary
variables have been used to train the predictor. This can lead to
unintended behavior unnoticed. Instead, the default should be the other
way around: only columns explicitly labeled as features are used as
such, all other columns are kept as auxiliary information.

A side-effect is that the instance attribute aux_vars is not needed
anymore. Also, the building id should be retained as an auxiliary
variable in addition to its use as an index.