Introduce model evaluation metrics for energy modeling
When using the predicted building age for energy modeling, the
prediction error propagates and results in deviations in terms of the
presumed energy demand of the building. To assess how suitable the
prediction is for an energy modeling use-case, the MAPE and R2 have been
introduces as additional evaluation metrics for age predictor models and
model comparisons.

The (heating) energy demand for certain residential building types is
given by TABULA (https://webtool.building-typology.eu/) and the building
types (SFH, MFH, TH and AB) estimated based on available information on
the building's footprint, floors, height and neighbors.