Allow to compare predictors across multiple seeds to improve robustness
The respective eval metrics will be averaged across different seeds to
get a more robust comparison result. Additionally, the standard
deviation across the different seeds is included in the results.

For some metrics, like spatial autocorrelation or feature importance
analysis multiple seeds are not yet supported and the result will only
be reported on the first seed.